% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_compare.R
\name{contrast_sig_genes}
\alias{contrast_sig_genes}
\title{Significant genes of a single contrast}
\usage{
contrast_sig_genes(
  res,
  padj_thr = 0.05,
  lfc_thr = 1,
  direction = c("either", "up", "down")
)
}
\arguments{
\item{res}{a DE results data.frame (validated)}

\item{padj_thr}{adjusted p-value threshold}

\item{lfc_thr}{absolute log2FoldChange threshold}

\item{direction}{one of "either" (default), "up", "down"}
}
\value{
a character vector of gene IDs passing the thresholds
}
\description{
Significant genes of a single contrast
}
