% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-count-phe.R
\name{create_count_phe}
\alias{create_count_phe}
\alias{create_count_phe.default}
\title{create_count_phe}
\usage{
create_count_phe(GSE = NULL)

\method{create_count_phe}{default}(GSE = NULL)
}
\arguments{
\item{GSE}{The number of GSE}
}
\value{
a list
}
\description{
This function would generate a list including a count_data and col_data as input data for testing
}
\examples{
input = create_count_phe()
}
