% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq-result.R
\name{define_de}
\alias{define_de}
\title{define_de}
\usage{
define_de(
  DEres,
  fold_change = "log2FoldChange",
  adj_pvalue = "padj",
  de_class = "regulate",
  fold_change_line = 1,
  adj_pvalue_line = 0.05
)
}
\arguments{
\item{DEres}{a data.frame contains fold_change, adj_p columns}

\item{fold_change}{a character, the name of column represented log2(fold change)}

\item{adj_pvalue}{a character, the name of column represented p-value}

\item{de_class}{a character, the name of created column represented differential expression}

\item{fold_change_line}{a numeric, the threshold to define significant fold change}

\item{adj_pvalue_line}{a numeric, the threshold to define significant p-value}
}
\value{
a dataframe
}
\description{
This function would create a new column to define the differential expression
}
