% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq-result.R
\name{DESeq_res}
\alias{DESeq_res}
\title{DESeqRes}
\usage{
DESeq_res(dds, fold_change_line = 1, adj_pvalue_line = 0.05, ...)
}
\arguments{
\item{dds}{a matrix with row name}

\item{fold_change_line}{a numeric, the threshold to define significant fold change}

\item{adj_pvalue_line}{a numeric, the threshold of significant log2(fold change)}

\item{...}{Aavailable arguments to be passed to results}
}
\value{
a dataframe
}
\description{
This function would get the results based on DESeqDataSet object
}
\examples{
\dontrun{
input = createCountPhe()
dds = DEseqObj(input[[1]],input[[2]],ref_level="untreated")
res = DESeqRes(dds)
}
}
