% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_state.R
\name{empty_project}
\alias{empty_project}
\title{Create an empty project state}
\usage{
empty_project(name = "untitled")
}
\arguments{
\item{name}{project name (used as default filename)}
}
\value{
a list with the canonical project structure
}
\description{
Create an empty project state
}
\keyword{internal}
