% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_embed.R
\name{fig_pca_3d}
\alias{fig_pca_3d}
\title{Interactive 3D PCA plot (PC1 / PC2 / PC3)}
\usage{
fig_pca_3d(
  counts_mat,
  metadata = NULL,
  n_top = 500,
  color_by = NULL,
  title = NULL
)
}
\arguments{
\item{counts_mat}{normalized counts matrix}

\item{metadata}{sample metadata (column 1 = sample)}

\item{n_top}{number of variable genes}

\item{color_by}{name of the metadata column to color by (or NULL)}

\item{title}{plot title}
}
\value{
a plotly scatter3d object
}
\description{
Interactive 3D PCA plot (PC1 / PC2 / PC3)
}
