% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_text.R
\name{generate_methods_text}
\alias{generate_methods_text}
\title{Generate a Methods paragraph for an analysis}
\usage{
generate_methods_text(project)
}
\arguments{
\item{project}{a project list (organism, contrasts, enrichment, wgcna)}
}
\value{
a single character string of prose
}
\description{
Generate a Methods paragraph for an analysis
}
