% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{get_orgdb}
\alias{get_orgdb}
\title{Get the OrgDb annotation object for an organism}
\usage{
get_orgdb(organism)
}
\arguments{
\item{organism}{one of "human", "mouse", "rat"}
}
\value{
the loaded \code{OrgDb} object
}
\description{
Get the OrgDb annotation object for an organism
}
\keyword{internal}
