% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_decoupler.R
\name{get_tf_network}
\alias{get_tf_network}
\title{Fetch a transcription-factor regulon network (CollecTRI)}
\usage{
get_tf_network(organism)
}
\arguments{
\item{organism}{one of "human", "mouse", "rat"}
}
\value{
a network data.frame (\code{source}, \code{target}, \code{mor})
}
\description{
Fetch a transcription-factor regulon network (CollecTRI)
}
