% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{guess_id_type}
\alias{guess_id_type}
\title{Guess the identifier type of a gene vector}
\usage{
guess_id_type(ids)
}
\arguments{
\item{ids}{character vector of gene identifiers}
}
\value{
one of "ensembl", "entrez", "symbol" (by majority vote)
}
\description{
Guess the identifier type of a gene vector
}
\keyword{internal}
