% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wgcna.R
\name{hub_genes}
\alias{hub_genes}
\title{Intramodular hub genes}
\usage{
hub_genes(wg, module, n = 20)
}
\arguments{
\item{wg}{the list returned by \code{\link[=run_wgcna]{run_wgcna()}}}

\item{module}{module color}

\item{n}{number of hub genes to return}
}
\value{
a data.frame (gene, kME) sorted by descending kME
}
\description{
Ranks the genes of a module by their module membership (signed kME).
}
