% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.R
\name{is_hex6}
\alias{is_hex6}
\title{Test whether a value is a valid 6-digit hex color}
\usage{
is_hex6(x)
}
\arguments{
\item{x}{value to test (typically a string)}
}
\value{
logical of length 1
}
\description{
Test whether a value is a valid 6-digit hex color
}
\keyword{internal}
