% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_app}
\alias{launch_app}
\title{Alias for \code{\link[=run_app]{run_app()}}}
\usage{
launch_app(port = NULL, launch_browser = TRUE)
}
\arguments{
\item{port}{port to launch on (default: random free port)}

\item{launch_browser}{if TRUE, open in default browser}
}
\description{
Alias for \code{\link[=run_app]{run_app()}}
}
