% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_state.R
\name{load_project}
\alias{load_project}
\title{Load a project from disk}
\usage{
load_project(path)
}
\arguments{
\item{path}{file path}
}
\value{
the project list
}
\description{
Load a project from disk
}
