% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_colors.R
\name{make_palette}
\alias{make_palette}
\title{Build an n-color palette}
\usage{
make_palette(name, n = 100)
}
\arguments{
\item{name}{palette name (from PALETTE_CHOICES or viridis family)}

\item{n}{number of colors to generate}
}
\value{
character vector of n hex colors
}
\description{
Build an n-color palette
}
\keyword{internal}
