% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data}
\alias{mod_data}
\alias{mod_data_ui}
\alias{mod_data_server}
\title{Data input module}
\usage{
mod_data_ui(id)

mod_data_server(id)
}
\arguments{
\item{id}{namespace ID}
}
\description{
Handles file upload (counts, metadata, pre-computed DE results) and
exposes validated reactive objects to other modules.
}
\details{
Returns a list of reactives: \code{counts()}, \code{metadata()}, \code{de_results()},
\code{organism()}.
}
