% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_volcano.R
\name{mod_volcano}
\alias{mod_volcano}
\alias{mod_volcano_ui}
\alias{mod_volcano_server}
\title{Volcano plot module}
\usage{
mod_volcano_ui(id)

mod_volcano_server(id, de_reactive)
}
\arguments{
\item{id}{namespace ID}

\item{de_reactive}{a reactive returning a DE results data.frame}
}
\description{
Shiny module wrapping \code{\link[=fig_volcano]{fig_volcano()}} and \code{\link[=fig_volcano_interactive]{fig_volcano_interactive()}},
with all the controls from the original app.
}
