% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{organism_info}
\alias{organism_info}
\title{Look up annotation resources for an organism}
\usage{
organism_info(organism)
}
\arguments{
\item{organism}{one of "human", "mouse", "rat" (case-insensitive)}
}
\value{
a list with \code{orgdb}, \code{species}, \code{kegg}, \code{reactome}, \code{taxon}
}
\description{
Look up annotation resources for an organism
}
\keyword{internal}
