% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallGSEA.R
\name{plot_overallGSEA}
\alias{plot_overallGSEA}
\title{plot_overallGSEA}
\usage{
plot_overallGSEA(gsea_res, x = "Description", y = "NES")
}
\arguments{
\item{gsea_res}{Default dataset to use for plot.}

\item{x}{Default x of aesthetic mappings to use for plot. Generally, it is Description for enrichment pathway.}

\item{y}{Default y of aesthetic mappings to use for plot. Generally, it is NES.}
}
\value{
a ggplot2 object
}
\description{
This function would generate bar plot for GSEA results
}
\examples{
\dontrun{
 data(geneList, package="DOSE")
 geneChangeList <- sort(geneList, decreasing = TRUE)
 ego <- gseGO(geneChangeList, ont="BP",OrgDb=org.Hs.eg.db,eps=0)
 ego <- setReadable(ego, OrgDb =org.Hs.eg.db)
 data <- ego[127:147,1:9]
 print(overallGSEA(data,"Description","NES"))
 }
}
