% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_palettes.R
\name{rnaflow_colors}
\alias{rnaflow_colors}
\title{Qualitative RNAflow colors}
\usage{
rnaflow_colors(n)
}
\arguments{
\item{n}{number of colors needed}
}
\value{
a character vector of \code{n} hex colors (interpolated if \code{n} exceeds
the base palette)
}
\description{
Qualitative RNAflow colors
}
\keyword{internal}
