% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_export.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Save a ggplot to disk}
\usage{
save_ggplot(p, file, fmt = c("pdf", "png", "tiff"), w = 8, h = 6, dpi = 300)
}
\arguments{
\item{p}{a ggplot object}

\item{file}{output path}

\item{fmt}{"pdf", "png" or "tiff"}

\item{w, h}{dimensions in inches}

\item{dpi}{resolution (used for PNG/TIFF)}
}
\description{
Save a ggplot to disk
}
