% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_export.R
\name{save_pheatmap}
\alias{save_pheatmap}
\title{Save a pheatmap to disk}
\usage{
save_pheatmap(ph, file, fmt = c("pdf", "png", "tiff"), w = 8, h = 8, dpi = 300)
}
\arguments{
\item{ph}{a pheatmap object}

\item{file}{output path}

\item{fmt}{"pdf", "png" or "tiff"}

\item{w, h}{dimensions in inches}

\item{dpi}{resolution}
}
\description{
Save a pheatmap to disk
}
