% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_state.R
\name{save_project}
\alias{save_project}
\title{Save a project to disk}
\usage{
save_project(project, path)
}
\arguments{
\item{project}{a project list (from \code{\link[=empty_project]{empty_project()}} or session)}

\item{path}{file path (will get .rnaflow.rds extension if missing)}
}
\value{
invisibly returns the path written to
}
\description{
Save a project to disk
}
