% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_palettes.R
\name{scale_fill_omics_c}
\alias{scale_fill_omics_c}
\alias{scale_color_omics_c}
\title{Continuous fill scale for scores (sequential)}
\usage{
scale_fill_omics_c(palette = "batlow", ...)

scale_color_omics_c(palette = "batlow", ...)
}
\arguments{
\item{palette}{scico palette name}

\item{...}{passed to \code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn()}}}
}
\description{
Continuous fill scale for scores (sequential)
}
\keyword{internal}
