% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{session_manifest}
\alias{session_manifest}
\title{Key package versions used in the session}
\usage{
session_manifest()
}
\value{
a data.frame (package, version) for packages that are installed
}
\description{
Key package versions used in the session
}
\keyword{internal}
