% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-Seurat.R
\name{Seurat_Reference}
\alias{Seurat_Reference}
\title{Convert Seurat object to Reference object}
\usage{
Seurat_Reference(
  seurat,
  slot = "counts",
  assay = "RNA",
  cell_type_name = "cell_type",
  nUMI_name = NULL
)
}
\arguments{
\item{seurat}{A Seurat class object.}

\item{slot}{Specific assay data to get or set. see \code{\link[SeuratObject:AssayData]{SeuratObject::GetAssayData()}} for detail information.}

\item{assay}{Specific assay to get data from or set data for; defaults to the default assay. see \code{\link[SeuratObject:GetTissueCoordinates]{SeuratObject::GetTissueCoordinates()}} for detail information.}

\item{cell_type_name}{A colname in \code{Seurat@meta.data}, which refer to the group or cell type.}

\item{nUMI_name}{Optional, a colname (by pixel barcode) in Seurat object, which refer to total counts or UMI's appearing at each pixel. If not provided, nUMI will be assumed to be the total counts appearing on each pixel.}
}
\description{
convert Seurat class to Reference class as the reference data for RCTD analysis using spacexr package
}
