% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-Seurat.R
\name{Seurat_SpatialRNA}
\alias{Seurat_SpatialRNA}
\title{Convert Seurat object to SpatialRNA object}
\usage{
Seurat_SpatialRNA(
  seurat,
  slot = "counts",
  assay = "Spatial",
  image = NULL,
  nUMI_name = "nCount_Spatial"
)
}
\arguments{
\item{seurat}{A Seurat class object.}

\item{slot}{Specific assay data to get or set. see \code{\link[SeuratObject:AssayData]{SeuratObject::GetAssayData()}} for detail information.}

\item{assay}{Specific assay to get data from or set data for; defaults to the default assay. see \code{\link[SeuratObject:GetTissueCoordinates]{SeuratObject::GetTissueCoordinates()}} for detail information.}

\item{image}{Name of SpatialImage object to get coordinates for; if NULL, will attempt to select an image automatically.  see \code{\link[SeuratObject:GetTissueCoordinates]{SeuratObject::GetTissueCoordinates()}} for detail information.}

\item{nUMI_name}{Optional, a colname (by pixel barcode) in Seurat object, which refer to total counts or UMI's appearing at each pixel. If not provided, nUMI will be assumed to be the total counts appearing on each pixel.}
}
\description{
Convert Seurat class to SpatialRNA class as the input data for RCTD analysis using spacexr package.
}
