% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{supported_organisms}
\alias{supported_organisms}
\title{Supported organism keywords}
\usage{
supported_organisms()
}
\value{
a character vector
}
\description{
Supported organism keywords
}
\keyword{internal}
