% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_theme.R
\name{theme_exploration}
\alias{theme_exploration}
\title{Exploration-mode ggplot2 theme}
\usage{
theme_exploration(base_size = 13)
}
\arguments{
\item{base_size}{base font size}
}
\value{
a ggplot2 theme object
}
\description{
Slightly larger fonts, friendly margins, designed for screen viewing
during data exploration. Matches the look of the original app.
}
