% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_widgets.R
\name{ui_color_picker}
\alias{ui_color_picker}
\title{Color picker with synchronized hex text input}
\usage{
ui_color_picker(id, lbl, default)
}
\arguments{
\item{id}{input ID (Shiny will see \code{input[[id]]} with the hex value)}

\item{lbl}{display label}

\item{default}{starting hex color}
}
\value{
a UI element
}
\description{
Color picker with synchronized hex text input
}
\keyword{internal}
