% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_widgets.R
\name{ui_export_bar}
\alias{ui_export_bar}
\title{Export toolbar (format + dimensions + download button)}
\usage{
ui_export_bar(prefix, def_w, def_h)
}
\arguments{
\item{prefix}{prefix used to build input IDs (\verb{<prefix>_fmt}, \verb{_w}, \verb{_h},
\verb{_dpi}, \verb{_dl})}

\item{def_w, def_h}{default dimensions (inches)}
}
\value{
a UI element
}
\description{
Export toolbar (format + dimensions + download button)
}
\keyword{internal}
