% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_widgets.R
\name{ui_section_title}
\alias{ui_section_title}
\title{Section title (small caps style)}
\usage{
ui_section_title(lbl)
}
\arguments{
\item{lbl}{text to display}
}
\description{
Section title (small caps style)
}
\keyword{internal}
