% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_widgets.R
\name{ui_slider_num}
\alias{ui_slider_num}
\title{Synchronized slider + numeric input pair}
\usage{
ui_slider_num(slider_id, num_id, lbl, mn, mx, val, stp)
}
\arguments{
\item{slider_id}{slider input ID}

\item{num_id}{numeric input ID}

\item{lbl}{label}

\item{mn, mx}{min/max}

\item{val}{initial value}

\item{stp}{step}
}
\value{
a UI element
}
\description{
Useful for fine control when a slider alone is too coarse.
}
\keyword{internal}
