% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_widgets.R
\name{ui_stat_badge}
\alias{ui_stat_badge}
\title{Stat badge (number + label, colored background)}
\usage{
ui_stat_badge(n, lbl, bg)
}
\description{
Stat badge (number + label, colored background)
}
\keyword{internal}
