% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_components.R
\name{ui_stat_tile}
\alias{ui_stat_tile}
\title{Coloured stat tile (big number + label)}
\usage{
ui_stat_tile(n, lbl, bg = "#1D9E75")
}
\arguments{
\item{n}{the number / value to show}

\item{lbl}{the caption under it}

\item{bg}{background colour (hex)}
}
\value{
a \code{div} tag
}
\description{
Coloured stat tile (big number + label)
}
\keyword{internal}
