% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{volcano}
\usage{
volcano(
  de_res,
  filename = NULL,
  output_dir = ".",
  fold_change = "log2FoldChange",
  adj_pvalue = "padj",
  de_class = "regulate",
  fold_change_line = 1,
  adj_pvalue_line = 0.05
)
}
\arguments{
\item{de_res}{a dataframe which would be generated by DESeqRes function}

\item{filename}{NULL or character, the prefix of output. When value is NULL, the figures are not saved.}

\item{output_dir}{the path of the directory to save plot}

\item{fold_change}{a character, the colname of log2(Fold Change) in de_res}

\item{adj_pvalue}{a character, the colname of adjusted P-value in de_res}

\item{de_class}{a character, the colname of differential expression type in de_res}

\item{fold_change_line}{a numeric, the threshold of significant log2(fold change)}

\item{adj_pvalue_line}{a numeric, the threshold of significant adjusted P-value}
}
\description{
This function would plot volcano based on DESeqRes output and save as pdf and png format
}
\examples{
\dontrun{
input = createCountPhe()
dds = create_DEseq(input[[1]],input[[2]],ref_level="untreated")
res = DESeq_res(dds)
volcano(res)
}
}
