% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_wgcna.R
\name{wgcna_datexpr}
\alias{wgcna_datexpr}
\title{Build the WGCNA expression matrix}
\usage{
wgcna_datexpr(counts_norm, n_genes = 3000)
}
\arguments{
\item{counts_norm}{normalized matrix (genes x samples)}

\item{n_genes}{number of top-variance genes to keep}
}
\value{
a samples x genes numeric matrix (\code{datExpr})
}
\description{
Selects the most variable genes and transposes to samples x genes.
}
