% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_decoupler.R
\name{get_pathway_network}
\alias{get_pathway_network}
\title{Fetch a pathway-responsive-gene network (PROGENy)}
\usage{
get_pathway_network(organism, top = 500)
}
\arguments{
\item{organism}{one of "human", "mouse", "rat"}

\item{top}{number of most responsive genes per pathway}
}
\value{
a network data.frame (\code{source}, \code{target}, \code{weight}, ...)
}
\description{
Fetch a pathway-responsive-gene network (PROGENy)
}
