% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_heatmap.R
\name{mod_heatmap}
\alias{mod_heatmap}
\alias{mod_heatmap_ui}
\alias{mod_heatmap_server}
\title{Heatmap module}
\usage{
mod_heatmap_ui(id)

mod_heatmap_server(id, de_reactive, counts_reactive, metadata_reactive)
}
\arguments{
\item{id}{namespace ID}

\item{de_reactive}{reactive for DE results}

\item{counts_reactive}{reactive for counts (raw or normalized)}

\item{metadata_reactive}{reactive for metadata}
}
\description{
Heatmap module
}
