% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pca.R
\name{mod_pca}
\alias{mod_pca}
\alias{mod_pca_ui}
\alias{mod_pca_server}
\title{PCA module}
\usage{
mod_pca_ui(id)

mod_pca_server(
  id,
  counts_reactive,
  metadata_reactive,
  contrast_params_reactive = NULL
)
}
\arguments{
\item{id}{namespace ID}

\item{counts_reactive}{reactive for counts matrix (normalized recommended)}

\item{metadata_reactive}{reactive for metadata}

\item{contrast_params_reactive}{optional reactive returning the active
contrast's parameter list (to enable "restrict to contrast")}
}
\description{
PCA module
}
